requirements.txt
Django==3.1.4
python-dateutil>==2.8.0
djangorestframework==3.12.2
mysql-connector==2.2.9
psycopg2-binary>==2.9.0
flake8>==3.9.0
django-extensions==3.1.0
djangorestframework-simplejwt==4.6.0
requests>==2.20.0
Pillow>==8.3.0
django-crontab>==0.7.0
celery==5.1.2
redis==3.5.3